Move toolbar children into Toolbar
